FIPS: The Fatigue Impairment Prediction Suite

The "Fatigue Impairment Prediction Suite" (FIPS) is currently under development and implemented in the R programming language. FIPS provides researchers and practitioners comprehensive set of functions for applying bio-mathematical models (BMMs) of fatigue. FIPS is under active development and implemented in the R programming language. FIPS provides a set of well-documented functions for transforming sleep and actigraphy data to formats required for applying BMMs, as well as a set of functions for simulating and interpreting BMMs with several kinds of models and customisable parameter settings.

Details

BMMs are a class of biological phenomenological models which are used to predict the neuro-behavioural outcomes of fatigue (e.g., alertness, performance) using sleep-wake history. These models are frequently applied by defence and industrial sectors to support system safety as part of broader fatigue management strategies. FIPS is the first open-source BMM framework enabling practitioners to inspect, validate, and ideally extend BMMs. Although there are several different implementations of BMMs, most have their roots in Borbély's (1982) two process model which models sleepiness/performance impairment as the sum of two processes: a circadian process and a homeostatic process.

Author

Maintainer: Michael David Wilson michael.d.wilson@curtin.edu.au

Authors:

Luke Strickland luke.strickland@curtin.edu.au

Timothy Ballard timothy.ballard@uq.edu.au