FIPS Simulation dispatcher

FIPS_simulate is used to apply a particular BMM simulation to a FIPS_df. It will dispatch to the selected model simulation type and then return the df with the fitted model columns added on as a FIPS_simulation object.

FIPS_simulate(FIPS_df, modeltype = NULL, pvec, model_formula = NULL)

Arguments

FIPS_df

A valid FIPS_df series that has not already been modelled

modeltype

String: either "TPM" (Three Process Model) or "unified".

pvec

Parameter vector (named list), see default pvecs for guidance.

model_formula

An optional formula describing how the time-varying processes predictors should be calculated for the predicted output. See details.

Value

a FIPS_simulation object

Details

If the formula argument is omitted, then default prediction values will be returned: KSS and alertness for TPM, and fatigue (i.e., lapses) for unified.

Formula Argument

The formula argument takes the form of an R formula expression (e.g., y ~ c + s).

The term of the left-hand side (e.g., y) defines the variable name

The term(s) on the right-hand side define how the variable is computed

All variables must be defined in your enviornment or in the FIPS_simulation output (e.g., s, c, l, w for "TPM").

Parameter vector and other variable arguments must be placed in I(expression) format. For example, fatigue ~ s + c + I(pvec[["KSS_beta"]]).