Simulate: Three Process Model

Simulates three process model over specified period. Default parameters (la through S0) are constants used in previous applications of the model.

TPM_simulate(pvec, dat)

Arguments

pvec

a vector of default parameters

dat

input dataframe (ensure this is in FIPS format)

Value

dataframe with simulated values - where fatigue middle is estimate (if no error terms)

Details

Access the modelled tibble directly by calling the object.

Parameters

S function (homeostatic process) la = low asymptote (default = 2.4) ha = high asymptote (default = 14.3) d = rate of decay in alertness when awake (default = -0.0353) g = rate of recovery in alertness when asleep (default = log((ha-14)/(ha-7.96))/8) bl = break level in alertness, time at which low pressure sleep kicks in (default = 12.2)

C function (24-circadian process) Cm = average level of process (i.e., mesor; default = 0) Ca = amplitude of process (default = 2.5) p = phase or time at which process reaches its peak (default = 16.8)

U function (12-hour circadian process) Um = average level of process (i.e., mesor; default = -0.5) Ua = amplitude of process (default = 0.5)

W function (sleep intertia process) Wc = Initial reduction in alertness at time of waking (default = -5.72) Wd = Rate of recovery of alertness (default = -1.51)

Regression equation for converting alertness to KSS fatigue ratings a = intercept (default = 10.6) b = coefficient (default = -0.6)

See also