Simulate: Unified Model

Runs a full simulation of the 'Unified Model'.

unified_simulate(pvec, dat)

Arguments

pvec

a vector of default parameters, see [unified_pvec]

dat

input dataframe (ensure this is a FIPS_df)

Value

simulated dataset complete

References

Rajdev, P., Thorsley, D., Rajaraman, S., Rupp, T. L., Wesensten, N. J., Balkin, T. J., \& Reifman, J. (2013). A unified mathematical model to quantify performance impairment for both chronic sleep restriction and total sleep deprivation. Journal of Theoretical Biology, 331, 66--77. https://doi.org/10.1016/j.jtbi.2013.04.013

Ramakrishnan, S., Wesensten, N. J., Balkin, T. J., \& Reifman, J. (2016). A Unified Model of Performance: Validation of its Predictions across Different Sleep/Wake Schedules. Sleep, 39(1), 249--262. https://doi.org/10.5665/sleep.5358

See also